move network usage data to its own api route
